fix: 추억 상세 페이지에서 댓글 컴포넌트 분리
