+++ src/components/Comments.css
@@ -1,0 +1,167 @@
+.comment-section {
+  max-width: 1560px;
+  margin: 40px auto;
+  font-family: Arial, sans-serif;
+}
+
+.comment-section h3 {
+  font-size: 24px;
+  margin-bottom: 20px;
+}
+
+.comment-button-container {
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+.comment-button {
+  background-color: #282828;
+  color: white;
+  padding: 10px 20px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+  transition: background-color 0.3s;
+}
+
+.comment-button:hover {
+  background-color: #444;
+}
+
+.comment-list {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.comment-item {
+  border-bottom: 1px solid #e0e0e0;
+  padding: 15px 0;
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+}
+
+.comment-content {
+  flex-grow: 1;
+}
+
+.comment-nickname {
+  font-weight: bold;
+  margin-right: 10px;
+}
+
+.comment-date {
+  color: #888;
+  font-size: 14px;
+}
+
+.comment-content p {
+  margin-top: 10px;
+  margin-bottom: 0;
+}
+
+.comment-actions {
+  display: flex;
+}
+
+.comment-edit,
+.comment-delete {
+  background: none;
+  border: none;
+  cursor: pointer;
+  font-size: 16px;
+  margin-left: 10px;
+  color: #888;
+  transition: color 0.3s;
+}
+
+.comment-edit:hover,
+.comment-delete:hover {
+  color: #282828;
+}
+
+.pagination {
+  display: flex;
+  justify-content: center;
+  margin-top: 20px;
+}
+
+.pagination button {
+  background: none;
+  border: 1px solid #282828;
+  padding: 5px 10px;
+  margin: 0 5px;
+  cursor: pointer;
+  transition: background-color 0.3s, color 0.3s;
+}
+
+.pagination button.active {
+  background-color: #282828;
+  color: white;
+}
+
+.pagination button:hover:not(:disabled) {
+  background-color: #282828;
+  color: white;
+}
+
+.pagination button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.pagination img {
+  width: 16px;
+  height: 16px;
+}
+
+/* Modal styles */
+.modal {
+  background-color: white;
+  padding: 20px;
+  border-radius: 5px;
+  max-width: 500px;
+  width: 100%;
+}
+
+.modal h2 {
+  margin-top: 0;
+  margin-bottom: 20px;
+}
+
+.modal label {
+  display: block;
+  margin-top: 10px;
+}
+
+.modal input,
+.modal textarea {
+  width: 100%;
+  padding: 8px;
+  margin-top: 5px;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+}
+
+.modal textarea {
+  height: 100px;
+  resize: vertical;
+}
+
+.modal button {
+  background-color: #282828;
+  color: white;
+  padding: 10px 20px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+  margin-top: 20px;
+  transition: background-color 0.3s;
+}
+
+.modal button:hover {
+  background-color: #444;
+}
