+++ src/pages/Memory/MemoryDetail.css
@@ -115,208 +115,6 @@
   line-height: 1.6;
 }
 
-.comment-button-container {
-  text-align: center;
-  margin-top: 120px;
-}
-
-.comment-button {
-  background-color: #282828;
-  width: 400px;
-  height: 50px;
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.comment-button:hover {
-  background-color: #444;
-}
-
-.comment-section {
-  max-width: 1560px;
-  margin: 40px auto;
-}
-
-.comment-list {
-  list-style: none;
-  padding: 0;
-  max-height: 400px;
-  overflow-y: auto;
-}
-
-.comment-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  border-bottom: 1px solid #ddd;
-  padding: 15px 0;
-}
-
-.comment-content {
-  max-width: 70%;
-  word-wrap: break-word;
-}
-
-.comment-nickname {
-  font-weight: bold;
-  margin-right: 10px;
-}
-
-.comment-date {
-  font-size: 12px;
-  color: #999;
-}
-
-.comment-actions {
-  display: flex;
-  align-items: center;
-}
-
-.comment-edit,
-.comment-delete {
-  background: none;
-  border: none;
-  color: #666;
-  cursor: pointer;
-  font-size: 14px;
-  margin-left: 10px;
-}
-
-.comment-edit:hover,
-.comment-delete:hover {
-  color: #333;
-}
-
-/* 댓글 리스트 스크롤 스타일 */
-.comment-list::-webkit-scrollbar {
-  width: 8px;
-}
-
-.comment-list::-webkit-scrollbar-thumb {
-  background-color: #ccc;
-  border-radius: 10px;
-}
-
-.comment-list::-webkit-scrollbar-track {
-  background-color: #f9f9f9;
-}
-
-/* 모달 스타일 추가 */
-.modal {
-  background-color: white;
-  padding: 20px;
-  border-radius: 5px;
-}
-
-.modal h2 {
-  margin-bottom: 20px;
-}
-
-.modal label {
-  display: block;
-  margin-top: 10px;
-}
-
-.modal input,
-.modal textarea {
-  width: 100%;
-  padding: 5px;
-  margin-top: 5px;
-}
-
-.modal-submit-btn {
-  background-color: #282828;
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  margin-top: 20px;
-}
-
-.modal-submit-btn:hover {
-  background-color: #444;
-}
-
-.modal-input {
-  width: 100%;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-  margin-top: 10px;
-  font-size: 16px;
-}
-
-.modal-error-message {
-  color: red;
-  font-size: 14px;
-  margin-top: 10px;
-}
-
-/* 댓글 등록 모달 관련 스타일 */
-.comment-input {
-  width: 100%;
-  padding: 10px;
-  margin-top: 10px;
-  font-size: 16px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-  resize: vertical;
-}
-
-.comment-nickname-input,
-.comment-password-input {
-  width: 100%;
-  padding: 10px;
-  margin-top: 10px;
-  font-size: 16px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-}
-
-.pagination {
-  display: flex;
-  justify-content: center;
-  margin-top: 20px;
-}
-
-.pagination button {
-  background: #fff;
-  border: 1px solid #333;
-  padding: 5px 10px;
-  margin: 0 5px;
-  cursor: pointer;
-  color: #333;
-  border-radius: 5px;
-  transition: background-color 0.3s, color 0.3s;
-}
-
-.pagination button img {
-  width: 16px;
-  height: 16px;
-}
-
-.pagination button.active {
-  background-color: #333;
-  color: #fff;
-}
-
-.pagination button:hover {
-  background-color: #333;
-  color: #fff;
-}
-
-.pagination button:disabled {
-  background-color: #f0f0f0;
-  color: #ccc;
-  border-color: #ccc;
-  cursor: default;
-}
-
 /* 비밀번호 확인 페이지 스타일 추가 */
 .password-check-page {
   max-width: 400px;
