+++ src/pages/Memory/MemoryDetail.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import axios from "axios";
 import Modal from "../../components/Modal";
+import Comments from "../../components/Comments";
 import "./MemoryDetail.css";
 
 function MemoryDetail() {
@@ -12,7 +13,6 @@ function MemoryDetail() {
   const [likeCount, setLikeCount] = useState(0);
   const [showEditModal, setShowEditModal] = useState(false);
   const [showDeleteModal, setShowDeleteModal] = useState(false);
-  const [showCommentModal, setShowCommentModal] = useState(false);
   const [password, setPassword] = useState("");
   const [isPasswordRequired, setIsPasswordRequired] = useState(false);
   const [errorMessage, setErrorMessage] = useState("");
@@ -23,14 +23,6 @@ function MemoryDetail() {
     location: "",
     moment: "",
   });
-  const [comments, setComments] = useState([]);
-  const [newComment, setNewComment] = useState({
-    nickname: "",
-    content: "",
-    commentPassword: "",
-  });
-  const [currentPage, setCurrentPage] = useState(1);
-  const commentsPerPage = 3;
 
   useEffect(() => {
     const fetchMemoryData = async () => {
@@ -51,10 +43,6 @@ function MemoryDetail() {
           location: memory.location,
           moment: memory.moment,
         });
-
-        // 댓글 데이터 가져오기
-        const commentsResponse = await axios.get(`/api/posts/${memoryId}/comments`);
-        setComments(commentsResponse.data);
       } catch (error) {
         console.error("Error fetching memory data:", error);
       }
@@ -74,8 +62,6 @@ function MemoryDetail() {
 
   const handleEditClick = () => setShowEditModal(true);
   const handleDeleteClick = () => setShowDeleteModal(true);
-  const handleCommentModalOpen = () => setShowCommentModal(true);
-  const handleCommentModalClose = () => setShowCommentModal(false);
 
   const handleEditSubmit = async () => {
     try {
@@ -112,41 +98,6 @@ function MemoryDetail() {
     } catch (error) {
       console.error("Error verifying password:", error);
       setErrorMessage("비밀번호 확인에 실패했습니다.");
-    }
-  };
-
-  // 댓글 등록 함수
-  const handleCommentSubmit = async () => {
-    try {
-      await axios.post(`/api/posts/${memoryId}/comments`, newComment);
-      alert("댓글이 성공적으로 등록되었습니다.");
-      setNewComment({
-        nickname: "",
-        content: "",
-        commentPassword: "",
-      });
-      const commentsResponse = await axios.get(`/api/posts/${memoryId}/comments`);
-      setComments(commentsResponse.data);
-      setShowCommentModal(false);
-    } catch (error) {
-      alert("댓글 등록에 실패했습니다.");
-    }
-  };
-
-  // 댓글 페이지네이션
-  const indexOfLastComment = currentPage * commentsPerPage;
-  const indexOfFirstComment = indexOfLastComment - commentsPerPage;
-  const currentComments = comments.slice(indexOfFirstComment, indexOfLastComment);
-
-  const handleNextPage = () => {
-    if (currentPage < Math.ceil(comments.length / commentsPerPage)) {
-      setCurrentPage(currentPage + 1);
-    }
-  };
-
-  const handlePreviousPage = () => {
-    if (currentPage > 1) {
-      setCurrentPage(currentPage - 1);
     }
   };
 
@@ -211,8 +162,6 @@ function MemoryDetail() {
           <span className="date">{memoryData.moment}</span>
           <span className="divider">·</span>
           <span className="like-count">🌸 {likeCount}</span>
-          <span className="divider">·</span>
-          <span className="comment-count">💬 {comments.length}</span>
         </div>
       </div>
 
@@ -225,52 +174,7 @@ function MemoryDetail() {
         ))}
       </div>
 
-      <div className="comment-button-container">
-        <button className="comment-button" onClick={handleCommentModalOpen}>
-          댓글 등록하기
-        </button>
-      </div>
-
-      <div className="comment-section">
-        <h3>댓글 {comments.length}</h3>
-        <ul className="comment-list">
-          {currentComments.map((comment) => (
-            <li key={comment.id} className="comment-item">
-              <div className="comment-content">
-                <span className="comment-nickname">{comment.nickname}</span>
-                <span className="comment-date">{comment.date}</span>
-                <p>{comment.content}</p>
-              </div>
-              <div className="comment-actions">
-                <button className="comment-edit">✏️</button>
-                <button className="comment-delete">🗑️</button>
-              </div>
-            </li>
-          ))}
-        </ul>
-
-        <div className="pagination">
-          <button className="prev-button" onClick={handlePreviousPage} disabled={currentPage === 1}>
-            <img src="/arrow-left.svg" alt="Previous" />
-          </button>
-
-          {[...Array(Math.ceil(comments.length / commentsPerPage)).keys()].map(
-            (num) => (
-              <button
-                key={num + 1}
-                onClick={() => setCurrentPage(num + 1)}
-                className={currentPage === num + 1 ? "active" : ""}
-              >
-                {num + 1}
-              </button>
-            )
-          )}
-
-          <button className="next-button" onClick={handleNextPage} disabled={currentPage === Math.ceil(comments.length / commentsPerPage)}>
-            <img src="/arrow-right.svg" alt="Next" />
-          </button>
-        </div>
-      </div>
+      <Comments memoryId={memoryId} />
 
       <Modal showModal={showEditModal} handleClose={() => setShowEditModal(false)}>
         <h2>게시글 수정</h2>
@@ -309,28 +213,6 @@ function MemoryDetail() {
           onChange={(e) => setPassword(e.target.value)}
         />
         <button onClick={handleDeleteSubmit}>삭제하기</button>
-      </Modal>
-
-      <Modal showModal={showCommentModal} handleClose={handleCommentModalClose}>
-        <h2>댓글 등록하기</h2>
-        <label>닉네임</label>
-        <input
-          type="text"
-          value={newComment.nickname}
-          onChange={(e) => setNewComment({ ...newComment, nickname: e.target.value })}
-        />
-        <label>댓글 내용</label>
-        <textarea
-          value={newComment.content}
-          onChange={(e) => setNewComment({ ...newComment, content: e.target.value })}
-        />
-        <label>비밀번호</label>
-        <input
-          type="password"
-          value={newComment.commentPassword}
-          onChange={(e) => setNewComment({ ...newComment, commentPassword: e.target.value })}
-        />
-        <button onClick={handleCommentSubmit}>등록</button>
       </Modal>
     </div>
   );
